# do not write length of values before writing the values in the index