Fixes https://github.com/druid-io/druid/issues/3070

Converting a segments having multiple sketch columns to new format reduce the size of `00000.smoosh` file from 815003685 bytes to 759427013 bytes. Ran the benchmark on `get()` and `indexOf()` methods of `BufferIndexed` class against the same segment on a machine with pinned CPU frequency.

Performance results with current code in master -

```
Benchmark                             Mode  Cnt      Score     Error  Units
BufferIndexedBenchmark.benchmarkV1  avgt   10  74077.418 ± 220.287  us/op
```

Performance results with the proposed code changes -

```
Benchmark                            Mode  Cnt      Score     Error  Units
BufferIndexedBenchmark.benchmarkV1  avgt   10  73473.873 ± 408.886  us/op
BufferIndexedBenchmark.benchmarkV2  avgt   10  73052.990 ± 334.913  us/op
```

These are the benchmark numbers for the static method `fromIterable()` in `GenericIndexed` class.

Current Code in master -

```
Benchmark                                      Mode  Cnt      Score    Error  Units
GenericIndexedBenchmark.benchmarkFromIterable  avgt   10  15492.342 ± 42.773  us/op
```

After proposed changes - 

```
Benchmark                                        Mode  Cnt      Score    Error  Units
GenericIndexedBenchmark.benchmarkFromIterable    avgt   10  15480.567 ± 48.199  us/op
GenericIndexedBenchmark.benchmarkFromIterableV2  avgt   10  13931.499 ± 53.123  us/op
```

The format used for writing can be configured using property `druid.indexer.genericIndexed.version` if it is set to `v2` the new format will be used and if set to `v1` or not set then the old format will be used. The current code can read GenericIndexed of both the formats.