@drcrallen 2016-06-06
This performance change is pretty small. Was the benchmark performed on a laptop?