@himanshug 2016-06-06
@drcrallen this PR is less about improving speed but more about reducing the size of serialized data produced by GenericIndexed and hence segment size.