@pjain1 2016-06-07
@drcrallen Yes, I ran this benchmark multiple times on a mac (and I was not browsing reddit at that time) and most of the times the performance change was small that lead me to think that it should not affect performance in production env. Regarding "pinned frequency governor machine" I have to read about it and then will try to run the benchmark on such machine.