@gianm 2016-06-08
Old readers cannot read the new format, right? In that case I'd prefer writers to use the old format by default, otherwise people can't have control over when they migrate (they might want to not switch right away to preserve ability to roll back or to keep using some older tools that read segments directly).