@pjain1 2016-06-26
@gianm updated the code to write by default in old format and if `druid.indexer.genericIndexed.version` i set to `v2` then the new format will be used.

@drcrallen ran the benchmark on a machine with pinned CPU frequency 

updated the master comment to reflect current changes